caibaogpt/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==0.28.0
PyPDF2==3.0.1
python-dotenv==1.0.0 